transitRouting: use the `useHistoryTracking` for routing form
Let the form use the `useHistoryTracking` hook to allow to track object
changes history. As the `TransitRouting` object is a simple one, without
ID, or DB representation, it is a good candidate to be the first to
migrate to the new form structure, using hooks instead of inheritance
for history tracking.

Also add an UndoRedoButtons widget to allow to undo/redo changes in the
form.