ScenarioComparisonPanel: refactor to function component
As part of a refactoring to move away from `GenericObject`s and
`ObjectWithHistory` whose history will rather be tracked outside the
class itself, using hook, the `ScenarioComparisonPanel` is turned into a
function component. The functionalities of the `ChangeEventForm`, from
which the class inherited are temporarily copied into this form and will
be replaced by a hook shortly.

Drop unused props, no props are required for this form.