confirmModal: Support React elements as children
This allows to put more complex elements than simple text in the confirm
modal (for example, adding a checkbox to not show a dialog again),
without having to go through the dangerously set html path.

The confirm modal still handles the title and buttons. The children
elements, if set, will only replace the text.