Refactor accessibility map polygon generation to use postgis functions
Our accessibility map generate was taking a log of time. They were using turf function in javascript
which are not the best performing options.

This change the code to store the result in a temporary postgis table and merge the circle via
St_Union. Function was added to a new model/db/geometryUtils file to make it obvious that it
require a running DB.

We also took the opportunity to reduce the circule precision and increase performance by
lowering the quad_segs parameter. A full analysis is available in comment in the code.

With these changes, we go from about 18s for a polygon generation to less than 3s. (For a point
near Polytechnique Montreal, with an STM GTFS.)

The unit test we redone with Claude. Integration with DB was also added.