batchRoute: handle results through a result visitor
fixes #1549

Define a `BatchRouteResultVisitor` interface that will visit each result
from the batch calculation and do something about it. The visitor can do
whatever it wants with the result and it does not necessarily involve
generating files.

Move the current result file generation to an implementation of this
visitor: the `BatchRouteFileResultVisitor` class, to generate the
requested files.

The `TrRoutingBatch` class now uses this visitor class to handle results
at the end of the calculation. It does not yet support any other result
handling, but it prepares the way.