jobs: support pausing/resuming executable jobs
fixes #1545

Add a status to the database type for 'paused' jobs

Better document the job statuses: what they mean and which statuses are
accessible from which other statuses.

In the backend, add `pause` and `resume` methods to the `ExecutableJob`
class to allow pausing and resuming a job. Only jobs 'pending' or
'inProgress' can be paused, and 'cancelled' or 'paused' jobs can be
resumed.

Update the `isCancelled` job callback to also return true for paused
jobs.

Add socket routes to pause and resume specific jobs

From the frontend, let running/pending jobs be paused and resumed.