db: make the readForLines schedule query more efficient
fixes #1529

The schedule collection subquery now receives the array of line ids for
which to get the collection so it can add this `where` to the inner
query of the collection. For a large network with many lines, like the
whole Montreal region for many time periods, this can reduce the query
time from 25 seconds to a few milliseconds.

Add sequential tests for the `readForLines` query.