jobs: allow jobs to type their specific internal_data field
The internal helps a job manage its internal state. Currently, there is
a single `checkpoint` field that allows to resume a `Job` at a specific
point. Batch route and accessibility map jobs make use of it.

But there may be additional internal data that a job may store, for
example the evolutionary algorithm network design Job may want to save
various states and/or calculated cost data for each generation. An
additional type gives more flexibility to the job type.