jobs: add a `parent_job_id` field to jobs
Add the `parent_job_id` field to the jobs table. This field is nullable
and a foreign key to the job's table `ìd` field. It allows to specify
child jobs for a parent.

By default, the `collection` function only returns the parent jobs, as
they are responsible for their children.