jobs: add a `parent_job_id` field to jobs
Add the `parent_job_id` field to the jobs table. This field is nullable
and a foreign key to the job's table `ìd` field. It allows to specify
child jobs for a parent.

By default, the `collection` function all the jobs. To get only parent jobs, you
can set the parent_id to null.

This also update the delete function to first call delete on all the child tasks first
and then clean up the task.