simulation: move method and algorithm types to common
and drop the registerAlgorithm and registerSimulationMethod
functionalities

fixes #1552

For the frontend to be able to configure them, the simulation method
types and configuration descriptors should be in the transition-common
package.

Also we drop the registration of algorithm and simulation methods. The
API of the simulation is not yet stable enough to support dynamically
adding methods for now. So we hard-code what is currently implemented in
Transition, with the intent of properly typing what is and simplify the
upcoming work in making the simulations available from the frontend.

Rename the `EvolutionAlgorithm*` classes and types to `EvolutionaryAlgorithm*`.