networkDesign: add types for evolutionary transit network design job
fixes #1553

Add types for the `evolutionaryTransitNetworkDesign` job. It copies the
types from the current `Simulation` class, except that the
routingAttributes will later be part of the simulation methods
instead of defined for the whole job.

This also adds the worker code to run the job in the
`TransitionWorkerPool` thread.

The result types are still to be determined.