OdTripSimulation: make the demand type match the batchRoute job
fixes #1560

The `OdTripSimulationOptions` type is update to better split the
attributes into sub attributes:

* `demandAttributes`: match the type of the `batchRoute` job, and adds a
  field for the sampleRatio of the demand to evaluate each scenario. It
  also requests an optional trip weight field, from the trip's attributes.

* `transitRoutingAttributes`: They used to be defined at the
  network design level, but they belong to the simulation method

* `evaluationOptions`: Specify the cost functions to use to compare OD
  trip results between simulations.

The `SimulationAlgorithmOptionsDescriptor` type is updated to support
nested options, but also 'custom' (for now, for example for files and
csv fields). The `OdTripSimulation` method is rendered temporarily
ineffective, as issues #1542, #1545 and #1541 and maybe #1397 need to be
addressed first.