transitNetworkDesign: add panel and list of jobs
part of #1426

Add the translations for the network design interface strings

Add the networkDesign section to the UI, disabled by default. We can
enable it by default later when documentation and everything is set to
work. In the meantime, on can simply add the following to their local
`config.js` file:

```
sections: {
    networkDesign: {
        enabled: true
    }
}
```

Advertise the transit network design panel in the contributions.

Add the panel component, which lists the available
`evolutionaryTransitNetworkDesign` and allow to replay them. Users can
also create new network design jobs from scratch, but for now the form
is empty.