Regroup TrRouting process management from batch job into a single class
Both TrRoutingBatch and TrAccessibilityMapBatch jobs were handling TrRouting lifecycle in
a slightly different way. This regroup this management in a single class, TrRoutingBatchManager. (I know
not a great name). We add an option definition in TrRoutingBatchJobParameter, since is not yet used, but
will be once we had option to save a cache in the specific jobs.

Also we update the notication text to mention that we only start one server instead of multiple. This was changed a
long time ago, but the text and variable names still reflected the fact that we had multiple servers.

No unit test as it would require mocking almost everything. It's a thin wrapper around TrRoutingProcessManager.
Tested that both jobs still work well.