csv: add classes and components to manage csv field mappings
A common use case is to upload a csv file and map its fields to some
specific object data, for example, latitude/longitude of points, time,
name, ids, etc. This class and component simplifies this use case.

To use these new types and components, a description of the fields to
map is required: see the `CsvFieldMappingDescriptor` type for details.

There are 3 types of field mappings:

* `single`: maps a given key to a field in the csv field
* `latLon`: needs a latitude and longitude field from the csv. A key of
  this type will result in 2 mappings: `${key}Lat` and `${key}Lon`. Such
  a field will also require a `projection` to be specified. There is one
  `projection` per set of descriptors.
* `time`: represents a time field. For each time field, 2 additional
  information should be provided by the user: `${key}Type` will specify
  if the time is `arrival` or `departure` and `${key}Format` specifies the
  format of the time field (for example `HMM`, `HH:MM`, etc).

An object, `CsvFileMapping`, handles the validation/error messages, csv file
and mapping updates.

The React component `GenericCsvImportAndMappingForm` takes a current
mapping, along with an array of field mapping descriptors and
automatically handles the file selection and mapping widget updates.
Components requiring csv file mapping capabilities can just make use of
this component and pass a `onUpdate` callback to be notified of changes.