executableJob: add a function to prepare inputFiles for creation
This function is added in an `ExecutableJobUtils.ts` file and
takes a FileConfig object and creates the proper configuration to let
the new job get the files from where it is.

If the file was uploaded, it fetches in the user's import directory. If
it was part of a previous job, it loads the job, makes sure the user is
allowed and fetches the file location for the desired file key.