batchRouting: use the new csv file mapping types and components
The `TransitOdDemandFromCsv` now inherits the `CsvFileMapping` class
instead of the previous csv file. It calls the constructor with the
desired mapping descriptors. The demand type is updated to be of type
`BatchRoutingOdDemandFromCsvAttributes`, which is a `csvFileAndMapping`
with the mappings typed accordingly. The backend routes are updated
with the new types.

The OdTripProvider's option type is updated with the new mapped fields,
but except for naming, it works as before, using the new mappings.

In the frontend, the `BatchAttributesSelection` is not required anymore,
and the `ConfigureDemandFromCsvForm` component can be replaced with the
`GenericCsvImportAndMappingForm`.