networkDesign: Make the config types not partial
The types should not be partial, in the backend, they are meant to be
complete. Only the frontend will have partial types to support entering
the configuration, so the frontend will build their own types
accordingly.

In some tests, the simulation data is cast as `any` as those queries
will be removed in upcoming PRs, to be replaced with something else.