networkDesign frontend: make it possible to configure a job
This adds the required forms and components in the frontend to configure
a transit network design job with evolutionary algorithm. Types are not
completely set, those using default values do not populate the
configuration data, even if visible in the interface.

Add backend routes to run and replay a transit network design job.