networkDesign: update types for the `OdTripSimulationMethod` type
Make the demand in line with the one from batchRoute calculations.
Separate the simulation in 3 fields: demand, routing parameters for
calculation and additional evaluation options to calculate the cost of a
scenario. Add descriptors for each of those methods and update tests
accordingly.

In the backend, update the simulation method with the new parameters. It
does not work anymore. Pseudo-code has been written to describe how the
simulation method will have to run.