HACK: Prohibit child jobs from being enqueued until we can managed the parent/child relation automatically
This is to make sure a child job create end up being executed at restart for example.
We have not yet implemented a mecanism to handle yeilding from the parent and let children execute and
notification to wake up the parent with children are done.
For now the parent will need to "execute" the children directly