Fix number of parallel calculation to do in batch routing/accessibility
In 67d921d, we reworked the way we computed the number of thread to run (number of parallel queries to the routing engines).
We assumed we could rely on the check on max number of CPU that was present in TrRoutingProcessManager, but the number
we calculated in TrRoutingBatchManager was the one we used to set the promise queue size.

This reinstate the min check on maxParallelCalculation that was present before directly in TrRoutingBatchManager