networkDesign backend: checkpoint generation data
After the scenarios are prepared for a generation, add a checkpoint for
the generation, saving the candidate chromosomes and scenarioId.

Upon resuming, if the previous generation is undefined and there is
checkoint data, recover the previous candidates, as well as their
corresponding scenarios from cache.

For now, the whole generation will be re-calculated, the candidates's
result are not yet saved as checkpoints.