csv mapping form: remove a dependency to useEffect
The parent should already be aware of the `csvFieldMapper` dependency
update, so the `onUpdate` does not need to be called when this prop is
updated. This effect is only to listen to the upload status update. If
the `csvFieldMapper` is part of the dependency, this may cause an update
loop depending on how the parent handles the value change.