csvFieldMapper: add a generic type for the field mappings
Also, the `getFieldMappings` and the `getCurrentFileAndMapping` only
return a value when the whole mapping is valid and the type can be
guaranteed (the expected is that required field mappings are set and
optional are optional in the type too).

The get a partial mapping during the filling process, use the
`getPartialFieldMappings` method.