workerpool: pause jobs if disk quota is reached
fixes #1616

When a job is set to start running, if the disk quota is reached,
instead of throwing an error and failing the Job, we simply pause it
with a statusMessage containing the reason for the pause, so the user
can clean up the jobs and resume the paused one when ready. This will
allow started jobs to resume at checkpoint if it was saved.