networkDesign simulation: sample OD trips file in simulation jobs
When simulating a scenario candidate using the OD trips approach, we
sample the original demand file to contain only a certain ratio of the
data.

For each selected record, a random time is selected as departure time
between 8 and 9 am. This time data is added to the demand parameters of
the simulation job, as it may not be present in the main file.

For now, each file row has a "sample ratio" probability of being
included in the final file, which means the file sample may not
represent exactly the sample ratio because of the randomness induced.