odTrip simulation: fixes after rebase on latest main
batch csv demand type has changed for the batch calculation job. We
create a similar demande file for the OdTripSimulationMethod type.

For sake of simplicity, for now, we keep the whole CsvFileAndMapper type
instead of just the field mapping, like in the batch calculation, as it
would require a different backend and frontend type with the options
descriptor. When we clean the option descriptor classes, we can make
sure the case is supported and the options descriptor type apply only to
the options configuration and allow for subsequent types in the backend.