Add continueOnMissingGeojson option to task
Adds a prompt to the 2_importAndProcessOsmDataToFiles task that allows points of interests with no geojson to be skipped instead of throwing an error.
As the import area increases, it is almost guaranteed to get multiple of these, making this necessary when importing data.
We made a partial fix for this before, however we only skipped some of the problematic function calls so that the task would pass with one particular geojson input.
It makes more sense to have one param control all of them equally instead of picking and choosing, however.
Fix: #1645