dep: remove `react-select-event` dependency
This packages was previously used in the the `InputMultiselect` test,
but this test was deactivated, so until we fix it and see if we still
need this, it is removed from the dependency list.

Add a dependency to `@testing-library/dom`, as it seems to be missing
from other packages. `react-select-event` was pulling it and makes test
pass. Without it, tests fail.

Also put the InputMultiselect test as `todo`.