batch schedule edit: close the edit window after save
Add a `onClose` prop to the `TransitScheduleBatchEdit` component to be
called when the edit form is closed. It receives an boolean argument, to
reset the selection list or nost.

Also, filter the lines by keeping their IDs so that if the line
collection is updated from the server after saving the data, the lines
are retrieved by their IDs and the list is displayed again instead of
disappearing because the objects are not the same anymore.