Add min and max to integer inputs
Adds a min and max parameter to InputString components of type number, preventing the arrows on the side from incrementing or decrementing the number beyond those bounds.
Adds a min bound of zero to most inputs to prevent a negative number from being entered, but also add custom min and max bounds to other inputs that already raise errors when the number is outside the bounds, adding an extra layer of security to those inputs.
Fix: #1664